+++ SCM.Application/Models/DTOs/Offers/OfferDTO.cs
@@ -5,11 +5,11 @@ namespace SCM.Application.Models.DTOs.Offers
 {
     public class OfferDTO
     {
-        public BigInteger Id { get; set; }
-        public BigInteger RequestId { get; set; }
+        public Int64 Id { get; set; }
+        public Int64 RequestId { get; set; }
         public decimal Amount { get; set; }
         public string SupplierName { get; set; }
-        public BigInteger SupplierId { get; set; }
+        public Int64 SupplierId { get; set; }
         public DateTime OfferDate { get; set; }
         public OfferStatus Status { get; set; }
     }
