+++ SCM.Application/Models/DTOs/Products/ProductDTO.cs
@@ -4,8 +4,8 @@ namespace SCM.Application.Models.DTOs.Products
 {
     public class ProductDTO
     {
-        public BigInteger Id { get; set; }
-        public BigInteger CategoryId { get; set; }
+        public Int64 Id { get; set; }
+        public Int64 CategoryId { get; set; }
         public string Name { get; set; }
         public string Detail { get; set; }
         public int UnitInStock { get; set; }
