+++ SCM.Application/Models/RequestModels/Invoice/CreateInvoiceVM.cs
@@ -4,7 +4,7 @@ namespace SCM.Application.Models.RequestModels.Invoice
 {
     public class CreateInvoiceVM
     {
-        public BigInteger RequestId { get; set; }
+        public Int64 RequestId { get; set; }
         public DateTime InvoiceDate { get; set; }
     }
 }
