+++ SCM.Application/Models/RequestModels/Offers/CreateOfferVM.cs
@@ -4,7 +4,7 @@ namespace SCM.Application.Models.RequestModels.Offers
 {
     public class CreateOfferVM
     {
-        public BigInteger RequestId { get; set; }
+        public Int64 RequestId { get; set; }
         public string SupplierName { get; set; }
 
         public decimal Amount { get; set; }
