+++ SCM.Application/Services/Abstractions/IOfferService.cs
@@ -6,10 +6,10 @@ namespace SCM.Application.Services.Abstractions
 {
     public interface IOfferService
     {
-        Task<Result<OfferDTO>> GetOfferByIdAsync(int offerId);
+        Task<Result<OfferDTO>> GetOfferByIdAsync(Int64 offerId);
         Task<Result<List<OfferDTO>>> GetOffersByRequest(GetAllOfferByRequestVM getAllOfferByRequestVM);
         Task<Result<bool>> CreateOfferAsync(CreateOfferVM createOfferVM);
         Task<Result<bool>> UpdateOfferAsync(UpdateOfferVM updateOfferVM);
-        Task<Result<bool>> DeleteOfferAsync(int offerId);
+        Task<Result<bool>> DeleteOfferAsync(Int64 offerId);
     }
 }
