+++ SCM.Domain/Entities/Product.cs
@@ -6,7 +6,7 @@ namespace SCM.Domain.Entities
     public class Product : AuditableEntity
     {   
 
-        public BigInteger CategoryId { get; set; }
+        public Int64 CategoryId { get; set; }
         public string Name { get; set; }
         public string Detail { get; set; }
         public int UnitInStock { get; set; }
